+++ test_muacrypt/test_cmdline.py
@@ -25,7 +25,7 @@ def test_init_and_make_header(mycmd):
         *AccountManager*not initialized*
     """)
     adr = "x@yz.org"
-    mycmd.run_ok(["add-account", "default", "--email-regex", adr])
+    mycmd.run_ok(["add-account", "--email-regex", adr])
     out = mycmd.run_ok(["make-header", adr])
     r = mime.parse_one_ac_header_from_string(out)
     assert "prefer-encrypt" not in out
@@ -42,7 +42,7 @@ def test_init_and_make_header_with_envvar(cmd, tmpdir):
 
 
 def test_exports_and_status_plain(mycmd):
-    mycmd.run_ok(["add-account", "default", "--email-regex=123@z.org"])
+    mycmd.run_ok(["add-account", "--email-regex=123@z.org"])
     out = mycmd.run_ok(["export-public-key"])
     check_ascii(out)
     out = mycmd.run_ok(["export-secret-key"])
@@ -70,7 +70,7 @@ def check_ascii(out):
 
 class TestProcessIncoming:
     def test_process_incoming(self, mycmd, datadir):
-        mycmd.run_ok(["add-account", "account1", "--email-regex=some@example.org"])
+        mycmd.run_ok(["add-account", "-a", "account1", "--email-regex=some@example.org"])
         mail = datadir.read("rsa2048-simple.eml")
         mycmd.run_fail(["process-incoming"], """
             *AccountNotFound*bob@testsuite.autocrypt.org*
@@ -90,10 +90,10 @@ class TestProcessIncoming:
         mycmd.run_ok(["status"])
 
     def test_process_incoming_no_autocrypt(self, mycmd, datadir):
-        mycmd.run_ok(["add-account", "account1", "--email-regex=b@b.org"])
+        mycmd.run_ok(["add-account", "--email-regex=b@b.org"])
         msg = mime.gen_mail_msg(From="Alice <a@a.org>", To=["b@b.org"], _dto=True)
         mycmd.run_ok(["process-incoming"], """
-            *processed*account1*no*Autocrypt*header*
+            *processed*default*no*Autocrypt*header*
         """, input=msg.as_string())
         mycmd.run_ok(["status"])
 
@@ -103,35 +103,43 @@ class TestAccountCommands:
         mycmd.run_ok(["status"], """
             *no accounts configured*
         """)
-        mycmd.run_ok(["add-account", "home", "--email-regex=home@example.org"], """
-            *account added*home*
+        mycmd.run_ok(["add-account", "--email-regex=home@example.org"], """
+            *account added*default*
         """)
         mycmd.run_ok(["status"], """
-            *account*home*
+            *account*default*
             *home@example.org*
         """)
-        mycmd.run_ok(["del-account", "home"])
+        mycmd.run_ok(["del-account"])
         mycmd.run_ok(["status"], """
             *no accounts configured*
         """)
 
+    def test_add_two_accounts_requires_option(self, mycmd):
+        mycmd.run_ok(["add-account"], """
+            *account added*default*
+        """)
+        mycmd.run_fail(["add-account"], """
+            AccountExists*default*
+        """)
+
     def test_modify_account_prefer_encrypt(self, mycmd):
-        mycmd.run_ok(["add-account", "default"])
+        mycmd.run_ok(["add-account"])
         mycmd.run_ok(["status"], """
             *account*default*
         """)
-        mycmd.run_ok(["mod-account", "default", "--prefer-encrypt=mutual"], """
+        mycmd.run_ok(["mod-account", "--prefer-encrypt=mutual"], """
             *account modified*default*
             *email?regex*.**
             *prefer-encrypt*mutual*
         """)
-        mycmd.run_ok(["mod-account", "default", "--email-regex=xyz"], """
+        mycmd.run_ok(["mod-account", "--email-regex=xyz"], """
             *account modified*default*
             *email?regex*xyz*
             *prefer-encrypt*mutual*
         """)
 
-        mycmd.run_ok(["mod-account", "default", "--prefer-encrypt=nopreference"], """
+        mycmd.run_ok(["mod-account", "--prefer-encrypt=nopreference"], """
             *account modified*default*
             *email?regex*xyz*
             *prefer-encrypt*nopreference*
@@ -141,7 +149,7 @@ class TestAccountCommands:
         adr = "x@y.org"
         keyhandle = bingpg.gen_secret_key(adr)
         monkeypatch.setenv("GNUPGHOME", bingpg.homedir)
-        mycmd.run_ok(["add-account", "home", "--use-key", adr,
+        mycmd.run_ok(["add-account", "--use-key", adr,
                       "--gpgbin=%s" % gpgpath, "--use-system-keyring"], """
                 *gpgmode*system*
                 *gpgbin*{}*
@@ -152,7 +160,7 @@ class TestAccountCommands:
         """)
 
     def test_test_email(self, mycmd):
-        mycmd.run_ok(["add-account", "home", "--email-regex=(home|office)@example.org"])
+        mycmd.run_ok(["add-account", "--email-regex=(home|office)@example.org"])
         mycmd.run_ok(["test-email", "home@example.org"])
         mycmd.run_ok(["test-email", "office@example.org"])
         mycmd.run_fail(["test-email", "xhome@example.org"], """
@@ -163,7 +171,7 @@ class TestAccountCommands:
 class TestProcessOutgoing:
 
     def test_simple(self, mycmd, gen_mail):
-        mycmd.run_ok(["add-account", "home"])
+        mycmd.run_ok(["add-account"])
         mail = gen_mail()
         out1 = mycmd.run_ok(["process-outgoing"], input=mail.as_string())
         m = mime.parse_message_from_string(out1)
@@ -175,7 +183,7 @@ class TestProcessOutgoing:
         assert x1 == x2
 
     def test_matching_account(self, mycmd, gen_mail):
-        mycmd.run_ok(["add-account", "account1", "--email-regex=account1@a.org"])
+        mycmd.run_ok(["add-account", "--email-regex=account1@a.org"])
         mail = gen_mail(From="x@y.org")
         # mycmd.run_fail(["process-outgoing"], input=mail.as_string(), fnl="""
         #     *AccountNotFound*x@y.org*
@@ -189,7 +197,7 @@ class TestProcessOutgoing:
         assert "account1@a.org" in msg2["Autocrypt"]
 
     def test_simple_dont_replace(self, mycmd, gen_mail):
-        mycmd.run_ok(["add-account", "account1"])
+        mycmd.run_ok(["add-account"])
         mail = gen_mail()
         gen_header = mycmd.run_ok(["make-header", "x@x.org"])
         mail.add_header("Autocrypt", gen_header)
@@ -203,7 +211,7 @@ class TestProcessOutgoing:
 
     @pytest.mark.parametrize("addr", ["a@a.org", "ño@example.org"])
     def test_sendmail(self, mycmd, gen_mail, popen_mock, addr):
-        mycmd.run_ok(["add-account", "account1"])
+        mycmd.run_ok(["add-account"])
         mail = gen_mail().as_string()
         pargs = ["-oi", addr]
         mycmd.run_ok(["sendmail", "-f", "--"] + pargs, input=mail)
@@ -217,8 +225,8 @@ class TestProcessOutgoing:
         assert "Autocrypt" in out_msg, out_msg.as_string()
 
     def test_sendmail_no_account(self, mycmd, gen_mail, popen_mock):
-        mycmd.run_ok(["add-account", "default", "--email-regex=account1@a.org"])
-        mycmd.run_ok(["mod-account", "default", "--email-regex", "123123"])
+        mycmd.run_ok(["add-account", "--email-regex=account1@a.org"])
+        mycmd.run_ok(["mod-account", "--email-regex", "123123"])
         mail = gen_mail().as_string()
         pargs = ["-oi", "b@b.org"]
         mycmd.run_fail(["sendmail", "-f", "--"] + pargs, input=mail)
@@ -232,7 +240,7 @@ class TestProcessOutgoing:
         # assert "Autocrypt" not in out_msg, out_msg.as_string()
 
     def test_sendmail_fails(self, mycmd, gen_mail, popen_mock):
-        mycmd.run_ok(["add-account", "default", "--email-regex=.*"])
+        mycmd.run_ok(["add-account", "--email-regex=.*"])
         mail = gen_mail().as_string()
         pargs = ["-oi", "b@b.org"]
         popen_mock.mock_next_call(ret=2)
@@ -246,7 +254,7 @@ class TestProcessOutgoing:
         assert "--qwe" in call.args
 
     def test_import_keydata(self, mycmd, datadir):
-        mycmd.run_ok(["add-account", "default"])
+        mycmd.run_ok(["add-account"])
         keydata = datadir.read_bytes("test1_autocrypt_org.key")
         mycmd.run_ok(["import-public-key"], input=keydata)
         out = mycmd.run_ok(["recommend", "test1@autocrypt.org"])
@@ -256,7 +264,7 @@ class TestProcessOutgoing:
         out = mycmd.run_ok(["recommend", "test1@autocrypt.org"])
         assert "available" in out
 
-        mycmd.run_ok(["mod-account", "default", "--prefer-encrypt", "mutual"])
+        mycmd.run_ok(["mod-account", "--prefer-encrypt", "mutual"])
         out = mycmd.run_ok(["import-public-key", "--prefer-encrypt=mutual"], input=keydata)
         assert "imported" in out
         out = mycmd.run_ok(["recommend", "test1@autocrypt.org"])
@@ -265,21 +273,21 @@ class TestProcessOutgoing:
 
 class TestRecommendation:
     def test_recommend_empty(self, mycmd):
-        mycmd.run_ok(["add-account", "home"])
+        mycmd.run_ok(["add-account", "-a", "home"])
         mycmd.run_ok(["recommend", "-a", "home", "unknown@email.org"])
 
     def test_recommend_one(self, mycmd):
         addr1 = "a@a.org"
         addr2 = "b@b.org"
-        mycmd.run_ok(["add-account", "ac1", "--email-regex", addr1])
-        mycmd.run_ok(["add-account", "ac2", "--email-regex", addr2])
+        mycmd.run_ok(["add-account", "-a", "ac1", "--email-regex", addr1])
+        mycmd.run_ok(["add-account", "-a", "ac2", "--email-regex", addr2])
 
         mycmd.send_mail(addr2, [addr1], Date=0)
         assert "available" == mycmd.parse_recommendation("ac1", [addr2])
 
         # switch addr2 and addr1 prefer_encrypt to "mutual", send a mail
-        mycmd.run_ok(["mod-account", "ac1", "--prefer-encrypt", "mutual"])
-        mycmd.run_ok(["mod-account", "ac2", "--prefer-encrypt", "mutual"])
+        mycmd.run_ok(["mod-account", "-a", "ac1", "--prefer-encrypt", "mutual"])
+        mycmd.run_ok(["mod-account", "-a", "ac2", "--prefer-encrypt", "mutual"])
         mycmd.send_mail(addr2, [addr1], Date=1)
         assert "encrypt" == mycmd.parse_recommendation("ac1", [addr2])
 
@@ -291,7 +299,7 @@ class TestRecommendation:
         addrs = []
         for i in range(1, 4):
             addr = "%d@x.org" % i
-            mycmd.run_ok(["add-account", "ac%d" % i, "--email-regex", addr])
+            mycmd.run_ok(["add-account", "-a", "ac%d" % i, "--email-regex", addr])
             addrs.append(addr)
         addr1, addr2, addr3 = addrs
 
@@ -301,7 +309,7 @@ class TestRecommendation:
 
         # switch all accounts to mutual
         for name in "ac1 ac2 ac2".split():
-            mycmd.run_ok(["mod-account", name, "--prefer-encrypt", "mutual"])
+            mycmd.run_ok(["mod-account", "-a", name, "--prefer-encrypt", "mutual"])
         mycmd.send_mail(addr2, [addr1])
         mycmd.send_mail(addr3, [addr1])
         assert "encrypt" == mycmd.parse_recommendation("ac1", [addr2, addr2])